requirements.txt
python-telegram-bot==20.7
Flask==2.3.3
openai==1.14.3
python-dotenv==1.0.1
aiofiles==23.2.1
